ai_logic/requirements.txt
google-cloud-vision>=3.0.0
google-generativeai>=0.8.0
pydantic>=2.0.0
requests>=2.28.0
Pillow>=10.0.0
pytesseract>=0.3.10
